+++ backend/src/main/java/com/MRSISA2021_T15/repository/MedicineRepository.java
@@ -14,7 +14,7 @@ public interface MedicineRepository extends CrudRepository<Medicine, Integer> {
 	Medicine findByMedicineCode(String medicineCode);
 	
 	@Query("select m.points from Medicine m where m.medicineCode like ?1")
-	double getPointsByMedicineCode(String medicineCode);
+	Integer getPointsByMedicineCode(String medicineCode);
 
 	@Query("select m from Medicine m where m.id = ?1")
 	public Medicine findMedicineWithId(Integer id);
